requirements.txt
numpy==1.21.2
pandas==1.3.2
requests==2.26.0
selenium==3.141.0
flask==2.0.2
psycopg==3.0.8
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
redis==4.1.1
Flask-Cors==3.0.10
flask-jwt==0.3.2
argon2-cffi==21.3.0

# Dev dependencies
factory-boy~=3.2.1
sqlalchemy2-stubs==0.0.2a19 #MyPy plugin